Split entity lookups from entitymanager
