+++ Robust.Client/Placement/Modes/AlignTileEmpty.cs
@@ -1,4 +1,5 @@
-﻿using Robust.Shared.Map;
+﻿using Robust.Shared.GameObjects;
+using Robust.Shared.Map;
 using Robust.Shared.Maths;
 
 namespace Robust.Client.Placement.Modes
@@ -51,7 +52,7 @@ namespace Robust.Client.Placement.Modes
             var topRight = new Vector2(CurrentTile.X + 0.99f, CurrentTile.Y + 0.99f);
             var box = new Box2(bottomLeft, topRight);
 
-            return !pManager.EntityManager.AnyEntitiesIntersecting(map, box);
+            return !EntitySystem.Get<SharedEntityLookupSystem>().AnyEntitiesIntersecting(map, box);
         }
     }
 }
