+++ Robust.Server/BaseServer.cs
@@ -67,7 +67,7 @@ namespace Robust.Server
 
         [Dependency] private readonly IConfigurationManagerInternal _config = default!;
         [Dependency] private readonly IComponentManager _components = default!;
-        [Dependency] private readonly IServerEntityManager _entities = default!;
+        [Dependency] private readonly IServerEntityManager _entityManager = default!;
         [Dependency] private readonly ILogManager _log = default!;
         [Dependency] private readonly IRobustSerializer _serializer = default!;
         [Dependency] private readonly IGameTiming _time = default!;
@@ -292,8 +292,6 @@ namespace Robust.Server
             IoCManager.Resolve<IGameTiming>().InSimulation = true;
 
             IoCManager.Resolve<INetConfigurationManager>().SetupNetworking();
-
-            _stateManager.Initialize();
             IoCManager.Resolve<IPlayerManager>().Initialize(MaxPlayers);
             _mapManager.Initialize();
             _mapManager.Startup();
@@ -303,8 +301,7 @@ namespace Robust.Server
 
             // Call Init in game assemblies.
             _modLoader.BroadcastRunLevel(ModRunLevel.Init);
-
-            _entities.Initialize();
+            _entityManager.Initialize();
 
             IoCManager.Resolve<ISerializationManager>().Initialize();
 
@@ -316,7 +313,8 @@ namespace Robust.Server
             prototypeManager.Resync();
 
             IoCManager.Resolve<IServerConsoleHost>().Initialize();
-            _entities.Startup();
+            _entityManager.Startup();
+            _stateManager.Initialize();
             _scriptHost.Initialize();
 
             _modLoader.BroadcastRunLevel(ModRunLevel.PostInit);
@@ -493,7 +491,7 @@ namespace Robust.Server
             _network.Shutdown($"Server shutting down: {_shutdownReason}");
 
             // shutdown entities
-            _entities.Shutdown();
+            _entityManager.Shutdown();
 
             if (_config.GetCVar(CVars.LogRuntimeLog))
             {
@@ -573,7 +571,7 @@ namespace Robust.Server
             }
 
             // Pass Histogram into the IEntityManager.Update so it can do more granular measuring.
-            _entities.TickUpdate(frameEventArgs.DeltaSeconds, TickUsage);
+            _entityManager.TickUpdate(frameEventArgs.DeltaSeconds, TickUsage);
 
             using (TickUsage.WithLabels("PostEngine").NewTimer())
             {
