+++ Robust.Server/GameStates/ServerGameStateManager.cs
@@ -48,6 +48,7 @@ namespace Robust.Server.GameStates
         /// <inheritdoc />
         public void Initialize()
         {
+            _entityView.Initialize();
             _networkManager.RegisterNetMessage<MsgState>(MsgState.NAME);
             _networkManager.RegisterNetMessage<MsgStateAck>(MsgStateAck.NAME, HandleStateAck);
 
@@ -188,7 +189,7 @@ namespace Robust.Server.GameStates
             var mailBag = _playerManager.GetAllPlayers()
                 .AsParallel()
                 .Where(s=>s.Status == SessionStatus.InGame).Select(GenerateMail);
-            
+
             foreach (var (msg, chan) in mailBag)
             {
                 // see session.Status != SessionStatus.InGame above
