+++ Robust.Server/Maps/MapLoader.cs
@@ -590,17 +590,7 @@ namespace Robust.Server.Maps
             {
                 foreach (var entity in Entities)
                 {
-                    _serverEntityManager.UpdateEntityTree(entity);
-                }
-
-                foreach (var entity in Entities)
-                {
                     _serverEntityManager.FinishEntityStartup(entity);
-                }
-
-                foreach (var entity in Entities)
-                {
-                    _serverEntityManager.UpdateEntityTree(entity);
                 }
             }
