+++ Robust.Shared/GameObjects/Components/Transform/TransformComponent.cs
@@ -1,13 +1,11 @@
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Linq;
 using Robust.Shared.Animations;
 using Robust.Shared.Containers;
 using Robust.Shared.IoC;
 using Robust.Shared.Map;
 using Robust.Shared.Maths;
-using Robust.Shared.Prototypes;
 using Robust.Shared.Players;
 using Robust.Shared.Serialization;
 using Robust.Shared.Serialization.Manager.Attributes;
@@ -125,7 +123,6 @@ namespace Robust.Shared.GameObjects
                 if (!DeferUpdates)
                 {
                     RebuildMatrices();
-                    UpdateEntityTree();
                     Owner.EntityManager.EventBus.RaiseEvent(
                         EventSource.Local, new RotateEvent(Owner, oldRotation, _localRotation));
                 }
@@ -286,8 +283,6 @@ namespace Robust.Shared.GameObjects
                         Owner.EntityManager.EventBus.RaiseEvent(
                             EventSource.Local, new MoveEvent(Owner, oldPosition, Coordinates));
                     }
-
-                    UpdateEntityTree();
                 }
                 else
                 {
@@ -319,7 +314,6 @@ namespace Robust.Shared.GameObjects
                 if (!DeferUpdates)
                 {
                     RebuildMatrices();
-                    UpdateEntityTree();
                     Owner.EntityManager.EventBus.RaiseEvent(
                         EventSource.Local, new MoveEvent(Owner, oldGridPos, Coordinates));
                 }
@@ -418,8 +412,6 @@ namespace Robust.Shared.GameObjects
                 // so duplicate additions (which will happen) don't matter.
                 ((TransformComponent) Parent!)._children.Add(Owner.Uid);
             }
-
-            UpdateEntityTree();
         }
 
         /// <inheritdoc />
@@ -430,7 +422,6 @@ namespace Robust.Shared.GameObjects
             // Keep the cached matrices in sync with the fields.
             RebuildMatrices();
             Dirty();
-            UpdateEntityTree();
         }
 
         public void RunDeferred(Box2 worldAABB)
@@ -443,7 +434,6 @@ namespace Robust.Shared.GameObjects
             }
 
             RebuildMatrices();
-            UpdateEntityTree(worldAABB);
 
             if (_oldCoords != null)
             {
@@ -579,7 +569,6 @@ namespace Robust.Shared.GameObjects
             SetPosition(newParent.InvWorldMatrix.Transform(_localPosition));
             RebuildMatrices();
             Dirty();
-            UpdateEntityTree();
         }
 
         internal void ChangeMapId(MapId newMapId)
@@ -613,11 +602,6 @@ namespace Robust.Shared.GameObjects
 
         private void MapIdChanged(MapId oldId)
         {
-            if (oldId != MapId.Nullspace)
-            {
-                Owner.EntityManager.RemoveFromEntityTree(Owner, oldId);
-            }
-
             Owner.EntityManager.EventBus.RaiseEvent(EventSource.Local, new EntMapIdChangedMessage(Owner, oldId));
         }
 
@@ -723,7 +707,6 @@ namespace Robust.Shared.GameObjects
                 }
 
                 Dirty();
-                UpdateEntityTree();
             }
 
             if (nextState is TransformComponentState nextTransform)
@@ -787,8 +770,6 @@ namespace Robust.Shared.GameObjects
 
             _invLocalMatrix = itransMat;
         }
-
-        public bool UpdateEntityTree(Box2? worldAABB = null) => Owner.EntityManager.UpdateEntityTree(Owner, worldAABB);
 
         public string GetDebugString()
         {
