+++ Robust.Shared/GameObjects/EntityManager.cs
@@ -28,7 +28,8 @@ namespace Robust.Shared.GameObjects
         [IoC.Dependency] private readonly IComponentFactory ComponentFactory = default!;
         [IoC.Dependency] private readonly IComponentManager _componentManager = default!;
         [IoC.Dependency] private readonly IGameTiming _gameTiming = default!;
-        [IoC.Dependency] private readonly IMapManager _mapManager = default!;
+
+        private SharedEntityLookupSystem _lookupSystem = default!;
 
         #endregion Dependencies
 
@@ -83,6 +84,7 @@ namespace Robust.Shared.GameObjects
 
         public virtual void Startup()
         {
+            _lookupSystem = EntitySystemManager.GetEntitySystem<SharedEntityLookupSystem>();
         }
 
         public virtual void Shutdown()
@@ -179,40 +181,6 @@ namespace Robust.Shared.GameObjects
             return query.Match(this);
         }
 
-        public IEnumerable<IEntity> GetEntitiesInMap(MapId mapId)
-        {
-            if (!_entityTreesPerMap.TryGetValue(mapId, out var trees))
-                yield break;
-
-            foreach (var entity in trees)
-            {
-                if (!entity.Deleted)
-                    yield return entity;
-            }
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesAt(MapId mapId, Vector2 position, bool approximate = false)
-        {
-            var list = new List<IEntity>();
-
-            var state = (list, position);
-
-            _entityTreesPerMap[mapId].QueryPoint(ref state, (ref (List<IEntity> list, Vector2 position) state, in IEntity ent) =>
-            {
-                var transform = ent.Transform;
-                if (MathHelper.CloseTo(transform.Coordinates.X, state.position.X) &&
-                    MathHelper.CloseTo(transform.Coordinates.Y, state.position.Y))
-                {
-                    state.list.Add(ent);
-                }
-
-                return true;
-            }, position, approximate);
-
-            return list;
-        }
-
         public IEnumerable<IEntity> GetEntities()
         {
             // Need to do an iterator loop to avoid issues with concurrent access.
@@ -282,13 +250,12 @@ namespace Robust.Shared.GameObjects
             EntityDeleted?.Invoke(this, entity.Uid);
             EventBus.RaiseEvent(EventSource.Local, new EntityDeletedMessage(entity));
         }
-        
+
         public void DeleteEntity(EntityUid uid)
         {
             if (TryGetEntity(uid, out var entity))
             {
                 DeleteEntity(entity);
-                UpdateEntityTree(entity);
             }
         }
 
@@ -412,6 +379,7 @@ namespace Robust.Shared.GameObjects
         private protected static void StartEntity(Entity entity)
         {
             entity.StartAllComponents();
+            EntitySystem.Get<SharedEntityLookupSystem>().UpdateEntityTree(entity);
         }
 
         private void CullDeletedEntities()
@@ -429,7 +397,7 @@ namespace Robust.Shared.GameObjects
 
                 AllEntities.RemoveSwap(i);
                 Entities.Remove(entity.Uid);
-                RemoveFromEntityTrees(entity);
+                _lookupSystem.RemoveFromEntityTrees(entity);
 
                 // Process the one we just swapped next.
                 i--;
@@ -464,277 +432,6 @@ namespace Robust.Shared.GameObjects
         /// Factory for generating a new EntityUid for an entity currently being created.
         /// </summary>
         protected abstract EntityUid GenerateEntityUid();
-
-#region Spatial Queries
-
-        /// <inheritdoc />
-        public bool AnyEntitiesIntersecting(MapId mapId, Box2 box, bool approximate = false)
-        {
-            var found = false;
-            _entityTreesPerMap[mapId].QueryAabb(ref found, (ref bool found, in IEntity ent) =>
-            {
-                if (!ent.Deleted)
-                {
-                    found = true;
-                    return false;
-                }
-
-                return true;
-            }, box, approximate);
-            return found;
-        }
-
-        /// <inheritdoc />
-        [MethodImpl(MethodImplOptions.AggressiveOptimization)]
-        public void FastEntitiesIntersecting(in MapId mapId, ref Box2 position, EntityQueryCallback callback)
-        {
-            if (mapId == MapId.Nullspace)
-                return;
-
-            _entityTreesPerMap[mapId]._b2Tree
-                .FastQuery(ref position, (ref IEntity data) => callback(data));
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesIntersecting(MapId mapId, Box2 position, bool approximate = false)
-        {
-            if (mapId == MapId.Nullspace)
-            {
-                return Enumerable.Empty<IEntity>();
-            }
-
-            var list = new List<IEntity>();
-
-            _entityTreesPerMap[mapId].QueryAabb(ref list, (ref List<IEntity> list, in IEntity ent) =>
-            {
-                if (!ent.Deleted)
-                {
-                    list.Add(ent);
-                }
-                return true;
-            }, position, approximate);
-
-            return list;
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesIntersecting(MapId mapId, Vector2 position, bool approximate = false)
-        {
-            const float range = .00001f / 2;
-            var aabb = new Box2(position, position).Enlarged(range);
-
-            if (mapId == MapId.Nullspace)
-            {
-                return Enumerable.Empty<IEntity>();
-            }
-
-            var list = new List<IEntity>();
-            var state = (list, position);
-
-            _entityTreesPerMap[mapId].QueryAabb(ref state, (ref (List<IEntity> list, Vector2 position) state, in IEntity ent) =>
-            {
-                if (Intersecting(ent, state.position))
-                {
-                    state.list.Add(ent);
-                }
-                return true;
-            }, aabb, approximate);
-
-            return list;
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesIntersecting(MapCoordinates position, bool approximate = false)
-        {
-            return GetEntitiesIntersecting(position.MapId, position.Position, approximate);
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesIntersecting(EntityCoordinates position, bool approximate = false)
-        {
-            var mapPos = position.ToMap(this);
-            return GetEntitiesIntersecting(mapPos.MapId, mapPos.Position, approximate);
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesIntersecting(IEntity entity, bool approximate = false)
-        {
-            if (entity.TryGetComponent<IPhysBody>(out var component))
-            {
-                return GetEntitiesIntersecting(entity.Transform.MapID, component.GetWorldAABB(), approximate);
-            }
-
-            return GetEntitiesIntersecting(entity.Transform.Coordinates, approximate);
-        }
-
-        /// <inheritdoc />
-        public bool IsIntersecting(IEntity entityOne, IEntity entityTwo)
-        {
-            var position = entityOne.Transform.MapPosition.Position;
-            return Intersecting(entityTwo, position);
-        }
-
-        private static bool Intersecting(IEntity entity, Vector2 mapPosition)
-        {
-            if (entity.TryGetComponent(out IPhysBody? component))
-            {
-                if (component.GetWorldAABB().Contains(mapPosition))
-                    return true;
-            }
-            else
-            {
-                var transform = entity.Transform;
-                var entPos = transform.WorldPosition;
-                if (MathHelper.CloseTo(entPos.X, mapPosition.X)
-                    && MathHelper.CloseTo(entPos.Y, mapPosition.Y))
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesInRange(EntityCoordinates position, float range, bool approximate = false)
-        {
-            var mapCoordinates = position.ToMap(this);
-            var mapPosition = mapCoordinates.Position;
-            var aabb = new Box2(mapPosition - new Vector2(range / 2, range / 2),
-                mapPosition + new Vector2(range / 2, range / 2));
-            return GetEntitiesIntersecting(mapCoordinates.MapId, aabb, approximate);
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesInRange(MapId mapId, Box2 box, float range, bool approximate = false)
-        {
-            var aabb = box.Enlarged(range);
-            return GetEntitiesIntersecting(mapId, aabb, approximate);
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesInRange(MapId mapId, Vector2 point, float range, bool approximate = false)
-        {
-            var aabb = new Box2(point, point).Enlarged(range);
-            return GetEntitiesIntersecting(mapId, aabb, approximate);
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesInRange(IEntity entity, float range, bool approximate = false)
-        {
-            if (entity.TryGetComponent<IPhysBody>(out var component))
-            {
-                return GetEntitiesInRange(entity.Transform.MapID, component.GetWorldAABB(), range, approximate);
-            }
-
-            return GetEntitiesInRange(entity.Transform.Coordinates, range, approximate);
-        }
-
-        /// <inheritdoc />
-        public IEnumerable<IEntity> GetEntitiesInArc(EntityCoordinates coordinates, float range, Angle direction,
-            float arcWidth, bool approximate = false)
-        {
-            var position = coordinates.ToMap(this).Position;
-
-            foreach (var entity in GetEntitiesInRange(coordinates, range * 2, approximate))
-            {
-                var angle = new Angle(entity.Transform.WorldPosition - position);
-                if (angle.Degrees < direction.Degrees + arcWidth / 2 &&
-                    angle.Degrees > direction.Degrees - arcWidth / 2)
-                    yield return entity;
-            }
-        }
-
-#endregion
-
-
-#region Entity DynamicTree
-
-        private readonly Dictionary<MapId, DynamicTree<IEntity>> _entityTreesPerMap =
-            new();
-
-        public virtual bool UpdateEntityTree(IEntity entity, Box2? worldAABB = null)
-        {
-            if (entity.Deleted)
-            {
-                RemoveFromEntityTrees(entity);
-                return true;
-            }
-
-            if (!entity.Initialized || !Entities.ContainsKey(entity.Uid))
-            {
-                return false;
-            }
-
-            var transform = entity.Transform;
-
-            DebugTools.Assert(transform.Initialized);
-
-            var mapId = transform.MapID;
-
-            if (!_entityTreesPerMap.TryGetValue(mapId, out var entTree))
-            {
-                entTree = new DynamicTree<IEntity>(
-                    GetWorldAabbFromEntity,
-                    capacity: 16,
-                    growthFunc: x => x == 16 ? 3840 : x + 256
-                );
-                _entityTreesPerMap.Add(mapId, entTree);
-            }
-
-            // for debugging
-            var necessary = 0;
-
-            if (entTree.AddOrUpdate(entity, worldAABB))
-            {
-                ++necessary;
-            }
-
-            foreach (var childTx in entity.Transform.ChildEntityUids)
-            {
-                if (UpdateEntityTree(GetEntity(childTx)))
-                {
-                    ++necessary;
-                }
-            }
-
-            return necessary > 0;
-        }
-
-        public bool RemoveFromEntityTree(IEntity entity, MapId mapId)
-        {
-            if (_entityTreesPerMap.TryGetValue(mapId, out var tree))
-            {
-                return tree.Remove(entity);
-            }
-
-            return false;
-        }
-
-        private void RemoveFromEntityTrees(IEntity entity)
-        {
-            foreach (var mapId in _mapManager.GetAllMapIds())
-            {
-                if (_entityTreesPerMap.TryGetValue(mapId, out var entTree))
-                {
-                    entTree.Remove(entity);
-                }
-            }
-        }
-
-        protected Box2 GetWorldAabbFromEntity(in IEntity ent)
-        {
-            if (ent.Deleted)
-                return new Box2(0, 0, 0, 0);
-
-            if (ent.TryGetComponent(out IPhysBody? collider))
-                return collider.GetWorldAABB(_mapManager);
-
-            var pos = ent.Transform.WorldPosition;
-            return new Box2(pos, pos);
-        }
-
-#endregion
     }
 
     public enum EntityMessageType : byte
