+++ Robust.Shared/GameObjects/IEntityManager.cs
@@ -83,13 +83,6 @@ namespace Robust.Shared.GameObjects
         IEnumerable<IEntity> GetEntities();
 
         /// <summary>
-        ///     Yields all of the entities on a particular map. faster than GetEntities()
-        /// </summary>
-        /// <param name="mapId"></param>
-        /// <returns></returns>
-        IEnumerable<IEntity> GetEntitiesInMap(MapId mapId);
-
-        /// <summary>
         /// Shuts-down and removes given <see cref="IEntity"/>. This is also broadcast to all clients.
         /// </summary>
         /// <param name="e">Entity to remove</param>
@@ -107,116 +100,5 @@ namespace Robust.Shared.GameObjects
         bool EntityExists(EntityUid uid);
 
         #endregion Entity Management
-
-        #region Spatial Queries
-
-        /// <summary>
-        /// Gets entities with a origin at the position.
-        /// </summary>
-        /// <param name="mapId"></param>
-        /// <param name="position"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        /// <returns></returns>
-        IEnumerable<IEntity> GetEntitiesAt(MapId mapId, Vector2 position, bool approximate = false);
-
-        /// <summary>
-        /// Checks if any entity is intersecting the box
-        /// </summary>
-        /// <param name="mapId"></param>
-        /// <param name="box"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        bool AnyEntitiesIntersecting(MapId mapId, Box2 box, bool approximate = false);
-        
-        void FastEntitiesIntersecting(in MapId mapId, ref Box2 position, EntityQueryCallback callback);
-
-        /// <summary>
-        /// Gets entities with a bounding box that intersects this box
-        /// </summary>
-        /// <param name="mapId"></param>
-        /// <param name="position"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        IEnumerable<IEntity> GetEntitiesIntersecting(MapId mapId, Box2 position, bool approximate = false);
-
-        /// <summary>
-        /// Gets entities with a bounding box that intersects this point
-        /// </summary>
-        /// <param name="mapId"></param>
-        /// <param name="position"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        IEnumerable<IEntity> GetEntitiesIntersecting(MapId mapId, Vector2 position, bool approximate = false);
-
-        /// <summary>
-        /// Gets entities with a bounding box that intersects this point
-        /// </summary>
-        /// <param name="position"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        IEnumerable<IEntity> GetEntitiesIntersecting(MapCoordinates position, bool approximate = false);
-
-        /// <summary>
-        /// Gets entities with a bounding box that intersects this point in coordinate form
-        /// </summary>
-        /// <param name="position"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        IEnumerable<IEntity> GetEntitiesIntersecting(EntityCoordinates position, bool approximate = false);
-
-        /// <summary>
-        /// Gets entities that intersect with this entity
-        /// </summary>
-        /// <param name="entity"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        IEnumerable<IEntity> GetEntitiesIntersecting(IEntity entity, bool approximate = false);
-
-        /// <summary>
-        /// Gets entities within a certain *square* range of this local coordinate
-        /// </summary>
-        /// <param name="position"></param>
-        /// <param name="range"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        IEnumerable<IEntity> GetEntitiesInRange(EntityCoordinates position, float range, bool approximate = false);
-
-        /// <summary>
-        /// Gets entities within a certain *square* range of this entity
-        /// </summary>
-        /// <param name="entity"></param>
-        /// <param name="range"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        IEnumerable<IEntity> GetEntitiesInRange(IEntity entity, float range, bool approximate = false);
-
-        /// <summary>
-        /// Gets whether two entities are intersecting each other
-        /// </summary>
-        /// <param name="entityOne"></param>
-        /// <param name="entityTwo"></param>
-        /// <returns></returns>
-        public bool IsIntersecting(IEntity entityOne, IEntity entityTwo);
-
-        /// <summary>
-        /// Gets entities within a certain *square* range of this bounding box
-        /// </summary>
-        /// <param name="mapID"></param>
-        /// <param name="box"></param>
-        /// <param name="range"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        IEnumerable<IEntity> GetEntitiesInRange(MapId mapID, Box2 box, float range, bool approximate = false);
-
-        /// <summary>
-        /// Get entities with bounding box in range of this whose center is within a certain directional arc, angle specifies center bisector of arc
-        /// </summary>
-        /// <param name="coordinates"></param>
-        /// <param name="range"></param>
-        /// <param name="direction"></param>
-        /// <param name="arcWidth"></param>
-        /// <param name="approximate">If true, will not recalculate precise entity AABBs, resulting in a perf increase. </param>
-        /// <returns></returns>
-        IEnumerable<IEntity> GetEntitiesInArc(EntityCoordinates coordinates, float range, Angle direction, float arcWidth, bool approximate = false);
-
-        #endregion
-
-        #region Spatial Updates
-
-        bool UpdateEntityTree(IEntity entity, Box2? worldAABB = null);
-        bool RemoveFromEntityTree(IEntity entity, MapId mapId);
-
-        #endregion
     }
 }
