+++ Robust.Shared/Timing/PauseManager.cs
@@ -14,7 +14,6 @@ namespace Robust.Shared.Timing
     internal sealed class PauseManager : IPauseManager, IPostInjectInit
     {
         [Dependency] private readonly IConsoleHost _conhost = default!;
-        [Dependency] private readonly IEntityManager _entityManager = default!;
         [Dependency] private readonly IMapManager _mapManager = default!;
 
         [ViewVariables] private readonly HashSet<MapId> _pausedMaps = new();
@@ -22,11 +21,13 @@ namespace Robust.Shared.Timing
 
         public void SetMapPaused(MapId mapId, bool paused)
         {
+            var lookupSystem = EntitySystem.Get<SharedEntityLookupSystem>();
+
             if (paused)
             {
                 _pausedMaps.Add(mapId);
 
-                foreach (var entity in _entityManager.GetEntitiesInMap(mapId))
+                foreach (var entity in lookupSystem.GetEntitiesInMap(mapId))
                 {
                     entity.Paused = true;
                 }
@@ -35,7 +36,7 @@ namespace Robust.Shared.Timing
             {
                 _pausedMaps.Remove(mapId);
 
-                foreach (var entity in _entityManager.GetEntitiesInMap(mapId))
+                foreach (var entity in lookupSystem.GetEntitiesInMap(mapId))
                 {
                     entity.Paused = false;
                 }
@@ -49,7 +50,7 @@ namespace Robust.Shared.Timing
 
             _unInitializedMaps.Remove(mapId);
 
-            foreach (var entity in _entityManager.GetEntitiesInMap(mapId))
+            foreach (var entity in EntitySystem.Get<SharedEntityLookupSystem>().GetEntitiesInMap(mapId))
             {
                 entity.RunMapInit();
                 entity.Paused = false;
@@ -65,7 +66,7 @@ namespace Robust.Shared.Timing
         {
             var mapId = _mapManager.GetGrid(gridId).ParentMapId;
 
-            foreach (var entity in _entityManager.GetEntitiesInMap(mapId))
+            foreach (var entity in EntitySystem.Get<SharedEntityLookupSystem>().GetEntitiesInMap(mapId))
             {
                 if (entity.Transform.GridID != gridId)
                     continue;
